services/extraction-agent/requirements.txt
Flask==3.0.0
requests==2.31.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
